+++ karbide-core/src/jvmAndAndroidMain/kotlin/dev/karmakrafts/karbide/BitUtils.kt
@@ -10,7 +10,7 @@ actual fun Short.reverseBits(count: Int): Short = (JInt.reverse(toInt()) ushr (I
 actual fun Int.reverseBits(count: Int): Int = JInt.reverse(this) ushr (Int.SIZE_BITS - count)
 actual fun Long.reverseBits(count: Int): Long = JLong.reverse(this) ushr (Long.SIZE_BITS - count)
 
-actual fun UByte.reverseBits(count: Int): UByte = reverseBitsCommon(count)
-actual fun UShort.reverseBits(count: Int): UShort = reverseBitsCommon(count)
+actual fun UByte.reverseBits(count: Int): UByte = toByte().reverseBits(count).toUByte()
+actual fun UShort.reverseBits(count: Int): UShort = toShort().reverseBits(count).toUShort()
 actual fun UInt.reverseBits(count: Int): UInt = toInt().reverseBits(count).toUInt()
 actual fun ULong.reverseBits(count: Int): ULong = toLong().reverseBits(count).toULong()